REV-25 Refactor App.js into index.js. Rename LoadingCircle to Loader. Remove unnecessary babel-polyfill.
